part 175 section 30 completed

+++ database/migrations/2022_05_06_201900_create_blog_post_tag_table.php
@@ -21,7 +21,7 @@ class CreateBlogPostTagTable extends Migration
                 ->on('blog_posts')
                 ->onDelete('cascade');
 
-            $table->unsignedInteger('tag_id')->index();
+            $table->unsignedBigInteger('tag_id')->index();
             $table->foreign('tag_id')->references('id')
                 ->on('tags')
                 ->onDelete('cascade');
